requirements.txt
arrow==1.2.3
attrs==22.2.0
bagit==1.8.1
bagit-profile==1.3.1
bdbag==1.6.3
boto3==1.26.123
botocore==1.29.123
certifi==2022.12.7
cfde-submit==0.2.5
cffi==1.15.1
chardet==5.1.0
charset-normalizer==3.0.1
click==7.1.1
colorama==0.4.6
commonmark==0.9.1
cryptography==39.0.0
d3b-utils==0.6.0.post2
decorator==5.1.1
fair-research-login==0.2.6
frictionless==4.16.2
gitdb==4.0.10
GitPython==3.1.30
globus-automate-client==0.12.2
globus-sdk==2.0.3
graphviz==0.12
greenlet==2.0.2
humanize==4.4.0
idna==3.4
isodate==0.6.1
Jinja2==3.1.2
jmespath==1.0.1
jsonschema==3.2.0
kf-utils==0.4.1
marko==1.2.2
MarkupSafe==2.1.1
modgrammar==0.10
numpy==1.24.3
osfclient==0.0.5
packaging==23.0
pandas==2.0.1
petl==1.7.12
psycopg2-binary==2.9.6
pycparser==2.21
Pygments==2.14.0
PyJWT==1.7.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
python-slugify==7.0.0
pytz==2022.7.1
PyYAML==5.4.1
requests==2.28.2
rfc3986==2.0.0
rich==9.13.0
s3transfer==0.6.0
setuptools-scm==5.0.2
shellingham==1.5.0.post1
simpleeval==0.9.12
six==1.16.0
smmap==5.0.0
SQLAlchemy==2.0.11
stringcase==1.2.0
tabulate==0.9.0
text-unidecode==1.3
tqdm==4.64.1
typer==0.3.2
typing-extensions==4.5.0
tzdata==2023.3
tzlocal==2.1
urllib3==1.26.14
validators==0.20.0
xmltodict==0.13.0
